Added main program to implement graph with satellites.  GETTING TONS OF ERRORS

+++ include/Edge.h
@@ -4,13 +4,13 @@
 struct Edge{
 
 	unsigned int destin;	// Location of Edge Destination
-	int weight;				// Edge Weight. Can be negative.
+	double weight;				// Edge Weight. Can be negative.
 
 	// Default Constructor
 	Edge() : destin(), weight( ) {}
 
 	// Overloaded Constructor
-	Edge( unsigned int dest, int weightIn ) : destin(dest), weight(weightIn) {}
+	Edge( unsigned int dest, double weightIn ) : destin(dest), weight(weightIn) {}
 
 };
